% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCamTrapData.R
\name{createCamTrapData}
\alias{createCamTrapData}
\title{import camera trapping data in the EcoDyn database}
\usage{
createCamTrapData(proj_name)
}
\description{
createCamTrapData
}
\details{
Function to import camera trapping data, including station sites, run times and link them to a project of the EcoDyn database.
}
