% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCamTrapSurvey.R
\name{createCamTrapSurvey}
\alias{createCamTrapSurvey}
\title{import camera trapping data in the EcoDyn database}
\usage{
createCamTrapSurvey(proj_name)
}
\description{
createCamTrapSurvey
}
\details{
Function to import camera trapping data, including station sites, run times and link them to a project of the EcoDyn database.
}
