% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createImages.R
\name{createImages}
\alias{createImages}
\title{add images from Camtrap folders to EcoDynDB}
\usage{
createImages()
}
\arguments{
\item{db_user}{Database user. User need to provide credentials with the right permissions.}
}
\description{
createImages
}
\details{
Function to read in camera trapping images from folders. The folder structure must correspond to the one created by camtrapR.
}
