% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPaper.R
\name{createPaper}
\alias{createPaper}
\title{create paper in the EcoDyn database}
\usage{
createPaper(db_user = NA)
}
\arguments{
\item{db_user}{Database user. User need to provide credentials with the right permissions.}
}
\description{
createPaper
}
\details{
Function to add a paper to the EcoDyn database.
}
