% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPcr.R
\name{createPcr}
\alias{createPcr}
\title{Get samples for new PCR}
\usage{
createPcr(out_dir = NA)
}
\arguments{
\item{out_dir}{Folder where the documentation files are saved.}
}
\description{
createPcr
}
\details{
Function to get list of unprocessed samples from the EcoDyn database and to calculate reagents of PCR. The relevant data is written into the EcoDyn database and some tables are updated.
}
