% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPoints.R
\name{createPoints}
\alias{createPoints}
\title{add geometries of points (e.g. camera-trapping stations)}
\usage{
createPoints(in_file = NA, ct_table = NA)
}
\arguments{
\item{db_user}{Database user. User need to provide credentials with the right permissions.}
}
\description{
createPoints
}
\details{
Function to read in geometries of points from tables.
}
