% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProject.R
\name{createProject}
\alias{createProject}
\title{create new project in the EcoDyn database}
\usage{
createProject(db_user = NA)
}
\arguments{
\item{db_user}{Database user. User need to provide credentials with the right permissions.}
}
\description{
createProject
}
\details{
Function to create a new project in the EcoDyn database.
}
