% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProjFolder.R
\name{createProjFolder}
\alias{createProjFolder}
\title{create new entry for project folder and folder content in the EcoDyn database}
\usage{
createProjFolder(proj_name)
}
\description{
createProjFolder
}
\details{
Function to create new entry for project folder and folder content in the EcoDyn database.
}
