% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProjKeywords.R
\name{createProjKeywords}
\alias{createProjKeywords}
\title{create new project keywords for the EcoDyn database}
\usage{
createProjKeywords(proj_name)
}
\description{
createProjKeywords
}
\details{
Function to create new project keywords for the EcoDyn database.
}
