% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProjOrg.R
\name{createProjOrg}
\alias{createProjOrg}
\title{enter new project organisations to the EcoDyn database}
\usage{
createProjOrg(proj_name)
}
\description{
createProjOrg
}
\details{
Function to add people to the EcoDyn database that participate in projects.
}
