% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProjPeople.r
\name{createProjPeople}
\alias{createProjPeople}
\title{enter new project participants to the EcoDyn database}
\usage{
createProjPeople()
}
\description{
createProjPeople
}
\details{
Function to add people to the EcoDyn database that participate in projects.
}
