% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSampleSheet.R
\name{createSampleSheet}
\alias{createSampleSheet}
\title{Create SampleSheet for Illumina MiSeq runs}
\usage{
createSampleSheet(out_dir = NA, pcr_batches = NULL)
}
\arguments{
\item{out_dir}{Folder containing the SampleSheets of a project. If not provided the user is requested select one.}

\item{pcr_batches}{Mandatory integers that characterizing the PCR batches that were quantified in the respective xls-files. For example "pcr_batches =c(1:5)" will query the database for the samples of the PCR batches "p001", "p002", "p003", "p004" and "p005", "pcr_batches =c(1, 25, 305)" will query for the PCR batches "p001", "p025" and "p305".}
}
\description{
createSampleSheet
}
\details{
Function to create standardized SampleSheets csv-files for Illumina MiSeq runs. The relevant data queried from the database and is written SampleSheet file.
}
