% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables.R
\name{db_tables}
\alias{db_tables}
\title{is_db_connected to EcoDyn database}
\usage{
db_tables()
}
\description{
is_db_connected
}
\details{
Creates an object 'db_structure' that shows the available schema and tables of the EcoDyn database.
}
