% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EcoDynConnect.r
\name{EcoDynConnect}
\alias{EcoDynConnect}
\title{connect to EcoDyn database}
\usage{
EcoDynConnect(
  db_user = NA,
  db_name = "EcoDynDB",
  db_host = "192.168.2.83",
  db_port = "5432"
)
}
\arguments{
\item{db_user}{Database user. User need to provide credentials with the right permissions.}

\item{db_name}{Database name. Must not be provided by the user.}

\item{db_host}{IP address of the server hosting the EcoDyn database. Must not be provided by the user.}

\item{db_port}{Server port of the EcoDyn database.}
}
\description{
EcoDynConnect
}
\details{
Function to connect to EcoDyn database. Deploys the DBI function dbConnect().
}
