% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EcoDynDisconnect.r
\name{EcoDynDisconnect}
\alias{EcoDynDisconnect}
\title{disconnect to EcoDyn database}
\usage{
EcoDynDisconnect()
}
\description{
EcoDynDisconnect
}
\details{
Function tests if a connection to the EcoDyn database exists and disconnect from it. Deploys the DBI function dbDisconnect().
}
