% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enterEx.R
\name{enterEx}
\alias{enterEx}
\title{Enter extraction}
\usage{
enterEx()
}
\description{
enterEx
}
\details{
Function to enter information on DNA extractions of samples from the EcoDynDB. The extr_names are created and data is written into the EcoDyn database.
}
