% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiPool.R
\name{equiPool}
\alias{equiPool}
\title{Paper sheets for the equimolar pooling}
\usage{
equiPool(out_dir = NA, pcr_batches = NULL)
}
\arguments{
\item{out_dir}{Output directory where the created files are saved.}

\item{pcr_batches}{Mandatory integers that characterizing the PCR batches that were quantified in the respective xls-files. For example "pcr_batches =c(1:5)" will query the database for the samples of the PCR batches "p001", "p002", "p003", "p004" and "p005", "pcr_batches =c(1, 25, 305)" will query for the PCR batches "p001", "p025" and "p305".}
}
\description{
equiPool
}
\details{
Create paper sheets for the equimolar pooling of sequencing libraries from PCR products.
}
