% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAssReads.R
\name{getAssReads}
\alias{getAssReads}
\title{create get assigned reads from the EcoDyn database}
\usage{
getAssReads()
}
\arguments{
\item{db_user}{Database user. User need to provide credentials with the right permissions.}
}
\description{
getAssReads
}
\details{
Function to queryread assigned reads from the EcoDyn database based on sample name, spies, genus or family and save them as fasta file.
}
