% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAssResults.R
\name{getAssResults}
\alias{getAssResults}
\title{get assignment results from the EcoDyn database}
\usage{
getAssResults()
}
\arguments{
\item{db_user}{Database user. User need to provide credentials with the right permissions.}
}
\description{
getAssResults
}
\details{
Function to read assignment results from the EcoDyn database and save them as xlsx file.
}
