% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLastPcrs.R
\name{getLastPcrs}
\alias{getLastPcrs}
\title{Get info on last PCR batches documented in the EcoDyn database}
\usage{
getLastPcrs(N = 5)
}
\arguments{
\item{N}{Number of PCRs listed}
}
\description{
getLastPcrs
}
\details{
Function to get list of documented PCR batches from the EcoDyn database.
}
