% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPcrInfo.r
\name{getPcrInfo}
\alias{getPcrInfo}
\title{Read in data of PCRs from standardized documentation files}
\usage{
getPcrInfo(in_dir = NA)
}
\arguments{
\item{in_dir}{Folder containing the PCR info xlsx-files. If not provided the user is requested select one.}
}
\description{
getPcrInfo
}
\details{
Function to read standardized PCR info xlsx-files. The relevant data is written into the EcoDyn database and some tables are updated.
}
