% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProjInfo.R
\name{getProjInfo}
\alias{getProjInfo}
\title{Read project realted information from EcoDyn database}
\usage{
getProjInfo()
}
\arguments{
\item{db_user}{Database user. User need to provide credentials with the right permissions.}
}
\description{
getProjInfo
}
\details{
Function to reads information like persons, organisations, etc. associated with a certain project from the database.
}
