% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeqRun.r
\name{getSeqRun}
\alias{getSeqRun}
\title{Read in data of sequencing runs from Illumina MiSeq runs}
\usage{
getSeqRun(in_dir = NA)
}
\arguments{
\item{in_dir}{Folder containing the PCR info xlsx-files. If not provided the user is requested select one.}
}
\description{
getSeqRun
}
\details{
Function to read relevant data from Illumina MiSeq runfolders. The relevant data is written into the EcoDyn database.
}
