% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isEcoDynConnected.r
\name{isEcoDynConnected}
\alias{isEcoDynConnected}
\title{test for connection to EcoDyn database}
\usage{
isEcoDynConnected()
}
\description{
isEcoDynConnected
}
\details{
Tests if an connection to the EcoDyn database exists. Deploys the DBI function dbIsValid().
}
