% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createEcoDynProject.R
\name{newEcoDynProject}
\alias{newEcoDynProject}
\title{create new project in the EcoDyn database}
\usage{
newEcoDynProject(db_user = NA)
}
\arguments{
\item{db_user}{Database user. User need to provide credentials with the right permissions.}
}
\description{
newEcoDynProject
}
\details{
Function to create a new project in the EcoDyn database.
}
