% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEcoDynTables.R
\name{readEcoDynTables}
\alias{readEcoDynTables}
\title{created overview of available EcoDyn database schemas and tables}
\usage{
readEcoDynTables()
}
\description{
readEcoDynTables
}
\details{
Creates an object 'db_structure' that shows the available schema and tables of the EcoDyn database.
}
