% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectProject.R
\name{selectProject}
\alias{selectProject}
\title{select project from the EcoDyn database}
\usage{
selectProject()
}
\description{
selectProject
}
\details{
Function to select project from the EcoDyn database.
}
