% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showPoints.R
\name{showPoints}
\alias{showPoints}
\title{view points of projects of the EcoDyn database}
\usage{
showPoints()
}
\description{
viewPoints
}
\details{
Function to view geometric points of projects of the EcoDyn database.
}
