% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateEcoDyn.R
\name{updateEcoDyn}
\alias{updateEcoDyn}
\title{Update sfb.number_of_pcrs table}
\usage{
updateEcoDyn()
}
\description{
updateEcoDyn
}
\details{
Function thta updates the sfb.number_of_pcrs table of EcoDynDB.
}
